add solution day 11 part2

+++ src/main/kotlin/solutions/day11monkeyinthemiddle/model/ThrowPolicy.kt
@@ -1,6 +1,6 @@
 package solutions.day11monkeyinthemiddle.model
 
-data class ThrowPolicy(val testDivisible : Int, val outcomeTrue : Int, val outcomeFalse : Int) {
+data class ThrowPolicy(val testDivisible : Long, val outcomeTrue : Int, val outcomeFalse : Int) {
     override fun toString(): String {
         return "if % $testDivisible == 0 send to $outcomeTrue else to $outcomeFalse"
     }
