+++ src/main/kotlin/solutions/day11monkeyinthemiddle/model/WorryAction.kt
@@ -2,13 +2,13 @@ package solutions.day11monkeyinthemiddle.model
 
 import solutions.day11monkeyinthemiddle.logic.WorryOperation
 
-class WorryAction(val worryOperation: WorryOperation, val argument : Int) {
+class WorryAction(val worryOperation: WorryOperation, val argument : Long) {
 
-    fun applyTo(input: Int) : Int {
+    fun applyTo(input: Long) : Long {
         return worryOperation.invoke(input, argument)
     }
 
-    fun applyTo(input: Int, newArgument : Int) : Int {
+    fun applyTo(input: Long, newArgument : Long) : Long {
         return worryOperation.invoke(input, newArgument)
     }
 
@@ -20,7 +20,7 @@ class WorryAction(val worryOperation: WorryOperation, val argument : Int) {
 
         fun from(s: String) : WorryAction {
             val worryOperation = WorryOperation.from(s)
-            val argument = Regex("[0-9]+").find(s)?.value?.toInt() ?: 0
+            val argument = Regex("[0-9]+").find(s)?.value?.toLong() ?: 0
             return WorryAction(worryOperation, argument)
         }
     }
